requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
email-validator
bcrypt==4.0.1
pg8000
asyncpg 
alembic 
pydantic-settings
